otel: Switch to global otel_setup
This is needed since it seems otel does not like use changing the
providers, so instead we have an session scoped autouse fixture that
sets things up with in memory implementations.

These are then used in captraces and capmetrics fixtures allowing tests
reliable access to the data.

While I'm here, I've also added headers to the get fixture.