oauth: Add detailed telemetry to token fetch logic
This makes sure to call record_exception() for each failed attempt, so
we get those events. And then sets the final error for the span at the
end.

Come to think of it, we should perhaps call set_status() as we go, and
instead mark OK at the end if it worked out. This would be more
defensive.