test: Rework OTelMocker and helpers
The mocker has been rewritten to avoid state, passing in the
exporters/readers and also not storing logs/metrics/spans.

Instead you always just get everything back, and instead use the
helpers.

All types are now wrapped in `Collected*` classes that provide access to
the bits we care about.

There are helpers for finding/getting items and asserting on them.